Fix gcc 4.0.3 warnings.
Try to fix gcc 3.2 warning about taking address of temporary.

+++ trunk/tu-testbed/gameswf/gameswf_sprite.cpp
@@ -160,10 +160,10 @@ namespace gameswf
 	{
 		if (is_enabled())
 		{
-			for (int i = 0; i < TU_ARRAYSIZE(FN_NAMES); i++)
+			for (size_t i = 0; i < TU_ARRAYSIZE(FN_NAMES); i++)
 			{
 				as_value dummy;
-				if (get_member(FN_NAMES[i], &dummy)) 
+				if (get_member(FN_NAMES[i], &dummy))
 				{
 					return true;
 				}
